Detect disconnected websocket connections
With the current implementation, a disconnected websocket is not
detected as *all* exceptions are silently swallowed, resulting in
`rtm_read()` returning an empty list forever without raising any
kind of exception.

This makes it impossible to detect that the websocket has been
disconnected and keeps client libraries reading forever while never
receiving data anymore.

With these changes, a `TimeoutError` will be thrown when the socket
gets disconnected (and other exceptions should bubble up as well) so
that users of this library know they must re-connect.