Allow not automatic retrying
Until something like tenacity usage merges it is not
easy to control the retrying loop that is happening internally
which can make shutting off the program that uses this client
hard (due to the usage of time.sleep).

To enable user's of this library to handle the retrying and
reconnecting externally (without the libraries help) just allow
for a 'auto_retry' keyword to be passed in to rtm_connect thaat
can disable (for now) the retrying.