Use a non-mutable default headers argument
When invoking `api_call()`, the default value for `headers` is a mutable
object! If at any point during invocation, `api_call` mutates this
dictionary, it will affect the behavior of future calls to the method.

Thankfully, the current source does not modify `headers`, so this
adverse behavior does not happen.

However, it's a much safer pattern (and consistent with the rest of this
file) to use `None` as the default argument, and construct a new
dictionary as required.

See: https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments