[#830] Clarify channel type in docs for Running the App
In issue #830, a developer tried to DM the bot and didn't
receive a response because the 'message.im' event was not used.

This tutorial only uses 'message.channels', so the bot must be
invited to a public channel to receive the "start" message.

This commit clarifies that the bot must be invited to a public channel.