Remove Dead Link From Readme
Hi Slackers,

I noticed that there was a dead anchor link in the readme table of contents for RTM client documentation. I understand RTM is deprecated (in favor of "socket mode") and a prior PR understandably deleted the introductory RTM documentation from the readme - https://github.com/slackapi/python-slack-sdk/pull/807. However, it seems this lone dead link to to the now-deleted content was left behind. My PR removes the dead link so nobody is confused trying to click it, like I was. :)

Thanks!