Update example code for rate limiting
The API methods themselves are inconsistent about the spelling of
ratelimited vs rate_limited. chat.postMessage spells it
rate_limited. The API docs on rate limiting itself suggest using the
HTTP status code.

This updates the examples in the doc to use the status_code member of
Response objects, instead of looking in the returned JSON.

Closes #1073