Per https://docs.python.org/3/howto/sockets.html
No data means the other side has closed the connection:

"When a recv returns 0 bytes, it means the other side has closed (or is in the
process of closing) the connection. You will not receive any more data on this
connection. Ever. You may be able to send data successfully"