rtm_v2 api spins in infinite loop under gevent (#1246)
* Raise an exception when reading from socket and getting no data.
Returning and empty bytes object causes an infinite loop between this and the
fetch function which in some cases can cause 100% cpu utilization (this happens
under gevent). Also the exception handling here can't just return an empty bytes for the same
reason - the now dead connection cannot be replaced by code that uses this
code.

* Per https://docs.python.org/3/howto/sockets.html
No data means the other side has closed the connection:
"When a recv returns 0 bytes, it means the other side has closed (or is in the
process of closing) the connection. You will not receive any more data on this
connection. Ever. You may be able to send data successfully"